Consolidated PGN read/write functions into core module.
