+++ chess_analyzer/core.py
@@ -2,8 +2,10 @@
 Functionality that's used in more than one module.
 """
 
+import re
 
-def parse_game_file(game_file):
+
+def parse_game_file(game):
     """
     Parses a string representation of a PNG file into a dictionary.
 
@@ -16,14 +18,28 @@ def parse_game_file(game_file):
     Returns:
         dict: A dictionary with all the PGN key-value pairs.
     """
+    key_re = re.compile(r"\[(?P<key>\S*)")
+    value_re = re.compile(r"\"(?P<value>.+)\"")
     parsed_game = {}
-    for item in game_file.split("\n"):
-        if item == "":
+    for row in game.split("\n"):
+        if row == "":
             continue
-        if item[0:2] == "1.":
-            parsed_game["game"] = item
+        elif row[0] == "[":
+            parsed_game[key_re.match(row).group("key").lower()] = value_re.search(
+                row
+            ).group("value")
         else:
-            item = item[1:-1]
-            split = item.find(" ")
-            parsed_game[item[0:split].lower()] = item[split + 2 : -1]
+            parsed_game["game"] = row
     return parsed_game
+
+
+def write_game_file(filename, game_file):
+    """
+    Write an output PGN game file.
+
+    Args:
+        filename (str): The output filename.
+        game_file (str): The PGN file contents.
+    """
+    with open(filename, "w") as f:
+        f.write(game_file)
